requirements.txt
matplotlib==3.2.2
numpy==1.19.5
Pillow==8.3.1
opencv-python==4.5.4.58
tensorflow==2.7.0
